+++ core/PhysiCell_utilities.h
@@ -80,8 +80,12 @@ namespace PhysiCell{
 
 long SeedRandom( long input );
 long SeedRandom( void );
+
 double UniformRandom( void );
 double NormalRandom( double mean, double standard_deviation );
+std::vector<double> UniformOnUnitSphere( void ); 
+std::vector<double> UniformOnUnitCircle( void ); 
+
 double dist_squared(std::vector<double> p1, std::vector<double> p2);
 double dist(std::vector<double> p1, std::vector<double> p2);